Moved prof_utils and mdir into the vcstools module
