+++ database/submit_to_database.py
@@ -23,13 +23,13 @@ import glob
 import textwrap as _textwrap
 
 #MWA software imports
-from vcstools import data_load
-
 from mwa_pulsar_client import client
+
+from vcstools import data_load
 from vcstools.metadb_utils import get_common_obs_metadata
 import find_pulsar_in_obs as fpio
 import sn_flux_est as snfe
-import prof_utils
+from vcstools import prof_utils
 
 
 import logging
