+++ setup.py
@@ -63,8 +63,8 @@ setup(name="mwa_vcstools",
                'scripts/plotting/plot_BPcal_128T.py',
                'database/submit_to_database.py', 'database/database_vcs.py',
                'utils/zapchan.py', 'utils/calc_ephem.py', 'utils/check_disk_usage.sh',
-               'utils/check_quota.sh', 'utils/mdir.py', 'utils/aocal.py', "utils/stickel.py",
-               'utils/sn_flux_est.py', 'utils/prof_utils.py',
+               'utils/check_quota.sh', 'utils/aocal.py', "utils/stickel.py",
+               'utils/sn_flux_est.py',
                'version.py'],
       setup_requires=['pytest-runner'],
       tests_require=['pytest']
