+++ vcstools/general_utils.py
